[IO] Add Json for file passing and serialisation (#18)
Add JSON output for the lexer.

+++ basic-v/source/Lexer/Tokeniser.hpp
@@ -1,3 +1,4 @@
+#include <memory>
 #include <string>
 #include <vector>
 
@@ -10,8 +11,8 @@ namespace bv::Lexer
     class Tokeniser
     {
     public:
-        static void TokeniseLines(std::vector<std::string>::iterator start, std::vector<std::string>::iterator end, int startLine, std::vector<bv::Token>* tokens);
+        static void TokeniseLines(std::vector<std::string>::iterator start, std::vector<std::string>::iterator end, uint64_t startLine, std::vector<bv::Token>* tokens);
 
-        static void TokeniseLine(std::string line, int lineNumber, std::vector<bv::Token>* tokens);
+        static void TokeniseLine(std::string line, uint64_t lineNumber, std::vector<bv::Token>* tokens);
     };
 }
